--- src/test/java/com/revature/controllerTest/CartControllerTest.java
@@ -1,65 +1,0 @@
-package com.revature.controllerTest;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.io.IOException;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.MethodOrderer;
-import org.junit.jupiter.api.Order;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestMethodOrder;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-
-import com.fasterxml.jackson.core.JsonParseException;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.revature.controllers.CartController;
-import com.revature.services.CartItemService;
-import com.revature.services.CartService;
-
-@ExtendWith(SpringExtension.class)
-@WebMvcTest(CartController.class)
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class CartControllerTest {
-	ObjectMapper om = new ObjectMapper();
-
-	@Autowired
-	CartController cartController;
-	
-    @Autowired
-	private MockMvc mockmvc;
-
-	@MockBean
-	private static CartService cservice;
-	@MockBean
-	private static CartItemService ciservice;
-	
-	public boolean isValidJSON(final String json) {
-		boolean valid = false;
-		try {
-			final JsonParser parser = new ObjectMapper().getFactory().createParser(json);
-			while (parser.nextToken() != null) {
-			}
-			valid = true;
-		} catch (JsonParseException jpe) {
-			jpe.printStackTrace();
-		} catch (IOException ioe) {
-			ioe.printStackTrace();
-		}
-
-		return valid;
-	}
-	@Test
-	@Order(1)
-	@DisplayName("1. AppContext Sanity Test")
-	public void contextLoads() throws Exception {
-
-		assertThat(cartController).isNotNull();
-	}
-}
